requirements.txt
# Flask Framework and Extensions
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
WTForms==3.0.1

# WSGI Server for Production
gunicorn==21.2.0

# Environment Variables Management
python-dotenv==1.0.0

# Machine Learning and Deep Learning
tensorflow==2.13.0
keras==2.13.1
numpy==1.24.3

# Image Processing
Pillow==10.0.0

# Utilities
Werkzeug==2.3.7

# Development Tools (optional)
coverage==7.3.0

# Note: Uncomment these if needed for deployment
# Fabric
# ecdsa
# paramiko
# pycrypto
